ClosetPage 컴포넌트 내의 필터링 관련 코드들을 다른 컴포넌트로 이동
- 필터링 관련 코드들이 이동된 컴포넌트: ButtonGroupsForFilteringFashionItems
- 컴포넌트가 마운트되었을 때, 전체 fashionItems가 화면에 보이지 않는 현상 있음
